@@ -1,6 +1,6 @@
 <?php
 /**
- * Copyright (c) 2013 Hawksearch (www.hawksearch.com) - All Rights Reserved
+ * Copyright (c) 2017 Hawksearch (www.hawksearch.com) - All Rights Reserved
  *
  * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
  * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
@@ -13,58 +13,57 @@
 
 namespace HawkSearch\Datafeed\Controller\Adminhtml\Hawkdatagenerate;
 
-use HawkSearch\Datafeed\Model\Datafeed;
+use Magento\Backend\App\Action;
 use Magento\Backend\App\Action\Context;
-use Magento\Framework\View\Result\PageFactory;
+use Magento\Framework\Controller\Result\JsonFactory;
+use HawkSearch\Datafeed\Helper\Data as Helper;
 
-class RunFeedGeneration
-    extends \Magento\Backend\App\Action
+class RunFeedGeneration extends Action
 {
     /**
-     * @var PageFactory
+     * @var JsonFactory
      */
-    protected $resultPageFactory;
-    private $helper;
+    protected $jsonResultFactory;
+
+    protected $helper;
 
     /**
-     * @param Context $context
-     * @param PageFactory $resultPageFactory
+     * @param Context     $context
+     * @param JsonFactory $jsonResultFactory
+     * @param Helper      $helper
      */
     public function __construct(
         Context $context,
-        PageFactory $resultPageFactory,
-        \HawkSearch\Datafeed\Helper\Data $helper
+        JsonFactory $jsonResultFactory,
+        Helper $helper
     ) {
         parent::__construct($context);
-        $this->resultPageFactory = $resultPageFactory;
+        $this->jsonResultFactory = $jsonResultFactory;
         $this->helper = $helper;
     }
 
     /**
-     * RunFeed execute
-     *
-     * @return void
+     * @return \Magento\Framework\App\ResponseInterface|\Magento\Framework\Controller\Result\Json|\Magento\Framework\Controller\ResultInterface
      */
     public function execute() {
-        $response = array('error' => false);
+        $result = $this->jsonResultFactory->create();
+        $data = ['error' => false];
         try {
             $disabledFuncs = explode(',', ini_get('disable_functions'));
             $isShellDisabled = is_array($disabledFuncs) ? in_array('shell_exec', $disabledFuncs) : true;
 
             if ($isShellDisabled) {
-                $response['error'] = 'This installation cannot run one off feed generation because the PHP function "shell_exec" has been disabled. Please use cron.';
+                $data['error'] = 'This installation cannot run one off feed generation because the PHP function "shell_exec" has been disabled. Please use cron.';
             } else {
-
                 if (strtolower($this->getRequest()->getParam('force')) == 'true') {
                     $this->helper->removeFeedLocks(true);
                 }
                 $this->helper->runDatafeed();
             }
         } catch (\Exception $e) {
-            $response['error'] = $e;
+            $data['error'] = $e;
         }
-        echo json_encode($response);
-
-        exit;
+        $result->setData($data);
+        return $result;
     }
 }
