@@ -10,86 +10,78 @@
  * FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS
  * IN THE SOFTWARE.
  */
+
 namespace HawkSearch\Datafeed\Model;
 class Email
 {
 
-/**
-* @var \Magento\Framework\Mail\Template\TransportBuilder
-*/
-protected $_transportBuilder;
- 
-/**
-* @var \Magento\Framework\Translate\Inline\StateInterface
-*/
-protected $inlineTranslation;
- 
-/**
-* @var \Magento\Framework\App\Config\ScopeConfigInterface
-*/
-protected $scopeConfig;
- 
-/**
-* @var \Magento\Store\Model\StoreManagerInterface
-*/
-protected $storeManager; 
-/**
-* @var \Magento\Framework\Escaper
-*/
-protected $_escaper;
-/**
-* @param \Magento\Framework\App\Action\Context $context
-* @param \Magento\Framework\Mail\Template\TransportBuilder $transportBuilder
-* @param \Magento\Framework\Translate\Inline\StateInterface $inlineTranslation
-* @param \Magento\Framework\App\Config\ScopeConfigInterface $scopeConfig
-* @param \Magento\Store\Model\StoreManagerInterface $storeManager
-*/
-public function __construct(
-\Magento\Framework\App\Action\Context $context,
-\Magento\Framework\Mail\Template\TransportBuilder $transportBuilder,
-\Magento\Framework\Translate\Inline\StateInterface $inlineTranslation,
-\Magento\Framework\App\Config\ScopeConfigInterface $scopeConfig,
-\Magento\Store\Model\StoreManagerInterface $storeManager,
-\Magento\Framework\Escaper $escaper
-) {
+    /**
+     * @var \Magento\Framework\Mail\Template\TransportBuilder
+     */
+    protected $_transportBuilder;
 
-$this->_transportBuilder = $transportBuilder;
-$this->inlineTranslation = $inlineTranslation;
-$this->scopeConfig = $scopeConfig;
-$this->storeManager = $storeManager;
-$this->_escaper = $escaper;
+    /**
+     * @var \Magento\Framework\Translate\Inline\StateInterface
+     */
+    protected $inlineTranslation;
+
+    /**
+     * @var \Magento\Framework\App\Config\ScopeConfigInterface
+     */
+    protected $scopeConfig;
+
+    /**
+     * @var \Magento\Store\Model\StoreManagerInterface
+     */
+    protected $storeManager;
+    /**
+     * @var \HawkSearch\Datafeed\Helper\Data
+     */
+    private $helper;
+
+    /**
+     * @param \Magento\Framework\App\Action\Context $context
+     * @param \Magento\Framework\Mail\Template\TransportBuilder $transportBuilder
+     * @param \Magento\Framework\Translate\Inline\StateInterface $inlineTranslation
+     * @param \Magento\Framework\App\Config\ScopeConfigInterface $scopeConfig
+     * @param \Magento\Store\Model\StoreManagerInterface $storeManager
+     * @param \HawkSearch\Datafeed\Helper\Data $helper
+     */
+    public function __construct(
+        \Magento\Framework\Mail\Template\TransportBuilder $transportBuilder,
+        \Magento\Framework\Translate\Inline\StateInterface $inlineTranslation,
+        \Magento\Framework\App\Config\ScopeConfigInterface $scopeConfig,
+        \Magento\Store\Model\StoreManagerInterface $storeManager,
+        \HawkSearch\Datafeed\Helper\Data $helper
+    )
+    {
+        $this->_transportBuilder = $transportBuilder;
+        $this->inlineTranslation = $inlineTranslation;
+        $this->scopeConfig = $scopeConfig;
+        $this->storeManager = $storeManager;
+        $this->helper = $helper;
+    }
+
+
+    public function sendEmail($templateParams)
+    {
+        $receiver = $this->helper->getCronEmail();
+        $this->inlineTranslation->suspend();
+        $sender = ['name' => $this->helper->getGenName(), 'email' => $this->helper->getGenEmail()];
+        $transport = $this->_transportBuilder
+            ->setTemplateIdentifier('hawksearch_datafeed_cronemail')
+            ->setTemplateOptions(
+                [
+                    'area' => \Magento\Framework\App\Area::AREA_ADMINHTML,
+                    'store' => \Magento\Store\Model\Store::DEFAULT_STORE_ID,
+                ]
+            )
+            ->setTemplateVars($templateParams)
+            ->setFrom($sender)
+            ->addTo($receiver)
+            ->getTransport();
+        $transport->sendMessage();
+
+        $this->inlineTranslation->resume();
+    }
 }
- 
-  
-  public function sendEmail($templateParams)
-  {
-	  
-
-	  	$objectManagerr = \Magento\Framework\App\ObjectManager::getInstance();	
-		$helper = $objectManagerr->create('HawkSearch\Datafeed\Helper\Data');
-		  
-		$receiver=$helper->getCronEmail();
-	    $this->inlineTranslation->suspend();
-	    $storeScope = \Magento\Store\Model\ScopeInterface::SCOPE_STORE; 
-	 // $sender = ['name' => $helper->getGenName(),'email' =>$helper->getGenEmail()];
-		$sender = ['name' => $helper->getGenName(),'email' =>$helper->getGenEmail()];
-		$transport = $this->_transportBuilder
-			->setTemplateIdentifier('hawksearch_datafeed_cronemail') 
-			->setTemplateOptions(
-				[
-				'area' => \Magento\Framework\App\Area::AREA_FRONTEND, 
-				'store' => \Magento\Store\Model\Store::DEFAULT_STORE_ID,
-			]
-			)
-			->setTemplateVars($templateParams)
-			->setFrom($sender)
-			->addTo($receiver)
-			->getTransport();
-		$transport->sendMessage();  
-
-		$this->inlineTranslation->resume();
-  }  
-  
-}
- 
-?>
